# BE/docs/swagger.yaml
definitions:
  auth.Login:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
...
info:
  contact: {}
paths:
  /v1/auth/login:
    post:
      description: Generating jwt Session Token
      parameters:
      - description: Payload
        in: body
        name: Payload
        required: true
        schema:
          $ref: '#/definitions/auth.Login'
      responses:
        "200":
          description: Ok
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "403":
          description: Forbidden
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "423":
          description: Locked
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Login User
      tags:
      - Auth
...
swagger: "2.0"